+++ src/main/java/javax/util/streamex/StreamEx.java
@@ -989,19 +989,40 @@ public class StreamEx<T> extends AbstractStreamEx<T, StreamEx<T>> {
         return sorted((Comparator<? super T>) Comparator.reverseOrder());
     }
 
+    /**
+     * Returns a {@code StreamEx} consisting of the distinct elements (according
+     * to {@link Object#equals(Object)}) which appear at least specified number
+     * of times in this stream.
+     *
+     * <p>
+     * This operation is not guaranteed to be stable: any of equal elements can
+     * be selected for the output. However if this stream is ordered then order
+     * is preserved.
+     *
+     * <p>
+     * This is a stateful quasi-intermediate operation.
+     *
+     * @param atLeast
+     *            minimal number of occurrences required to select the element.
+     *            If atLeast is 1 or less, then this method is equivalent to
+     *            {@link #distinct()}.
+     * @return the new stream
+     * @see #distinct()
+     * @since 0.3.1
+     */
     public StreamEx<T> distinct(long atLeast) {
         if (atLeast <= 1)
             return distinct();
         Spliterator<T> spliterator = stream.spliterator();
         Spliterator<T> result;
-        if(spliterator.hasCharacteristics(Spliterator.DISTINCT))
+        if (spliterator.hasCharacteristics(Spliterator.DISTINCT))
             // already distinct: cannot have any repeating elements
             result = Spliterators.emptySpliterator();
         else
             result = new DistinctSpliterator<>(spliterator, atLeast);
         return strategy().newStreamEx(StreamSupport.stream(result, stream.isParallel()).onClose(stream::close));
     }
-    
+
     /**
      * Returns a stream consisting of the results of applying the given function
      * to the every adjacent pair of elements of this stream.
@@ -1125,7 +1146,7 @@ public class StreamEx<T> extends AbstractStreamEx<T, StreamEx<T>> {
             List<T> res = new ArrayList<>();
             res.add(t);
             return res;
-        }).collapse((a, b) -> sameGroup.test(a.get(a.size()-1), b.get(0)), (a, b) -> {
+        }).collapse((a, b) -> sameGroup.test(a.get(a.size() - 1), b.get(0)), (a, b) -> {
             a.addAll(b);
             return a;
         });